pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ChatDKU"
version = "2.0.0"
authors = [
    { name="Mingxi Li" },
    { name="Yuxiang Lin", email="theta.lin@yandex.com" },
    { name="Cody Qin" },
    { name="Ningyuan Yang" },
    { name="Temuulen Enkhtamir" },
    { name="Anar Nyambayar" },
    { name="Munish Lohani" },
    { name="Sean Allen Siegfreid R. Bugarin" },
    { name="Zhiwei Li" },
    { name="Youran Wang" },
    { name="Zeyu Yu" },
    { name="Jingxuan Lin" },
    { name="Ruihan Yin" },
    { name="Haroon Butt" },
    { name="Halfi Khan" },
    { name="Kurtis Kwan" },
]
description = ""
readme = ""
requires-python = "<3.13, >=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    # core
    "llama-index~=0.14.20",
    "llama-index-storage-docstore-redis~=0.5.0",
    "llama-index-vector-stores-redis~=0.8.0",
    "llama-index-embeddings-text-embeddings-inference~=0.5.0",
    "llama-index-llms-llama-cpp~=0.6.0",
    "llama-index-retrievers-bm25~=0.7.1",
    "llama-index-readers-file~=0.6.0",
    "nltk>=3.9",
    "chromadb~=1.0.15",
    "redis~=5.2.1",
    "arize-phoenix~=13.21.0",
    "arize-phoenix-evals==2.13.0",
    "opentelemetry-api~=1.41.0",
    "opentelemetry-sdk~=1.41.0",
    "openinference-instrumentation~=0.1.46",
    "openinference-instrumentation-dspy~=0.1.34",
    "openinference-instrumentation-litellm~=0.1.30",
    "dspy~=3.1.3",
    "tokenizers>=0.21.0",
    "docx2txt",
    "python-pptx",
    "pymupdf", # for major ingestion
    "pymupdf4llm", # needed for pdf -> md
    "thefuzz", # for fuzzy searching major names in major ret
    "python-docx",
    "pdfplumber",
    "psycopg2-binary>=2.9.10",
    "pandas~=2.2.3",
    "dotenv>=0.9.9",
    "sqlalchemy",
    # Development dependencies
    "pytest",
    "pre-commit",
    # TUI
    "textual>=0.83.0",
    "pyfiglet>=1.0.2",
    "thefuzz>=0.22.1",
]

[project.optional-dependencies]
backend = [
    "eventlet~=0.39.1",
    "gunicorn~=23.0.0",
    "Django~=5.2.3",
    "django-appconf~=1.1.0",
    "django-cors-headers~=4.7.0",
    "django-cryptography~=1.1",
    "django-encrypted-model-fields~=0.6.5",
    "django-fernet-fields~=0.6",
    "django-import-export~=4.3.8",
    "djangorestframework~=3.16.0",
    "django-celery-beat~=2.8.1",
    "django_celery_results~=2.6.0",
    "celery~=5.5.3",
    "django-redis~=6.0.0",
    "openpyxl>=3.1.5",
    "locust>=2.39.0",
    "drf-spectacular[sidecar]",
    "django-prometheus",
]

evaluation = [
    "seaborn",
    "ray"
]

[tool.setuptools]
packages = ["chatdku"]

[tool.flake8]
ignore = ["E203","W503"]
max-line-length = 120

[dependency-groups]
dev = [
    "flake8>=7.3.0",
]
